Moved series-pass-notice to flexible tickets directory
