+++ src/Tickets/Blocks/Tickets/app/editor/not-supported-message/template.js
@@ -9,28 +9,21 @@ import PropTypes from 'prop-types';
  */
 import { __ } from '@wordpress/i18n';
 
-
-const NotSupportedMessage = ({
-	isBlockSelected,
-}) => {
+const NotSupportedMessage = () => {
 	return (
-		isBlockSelected
-		? (
-			<div className="tribe-editor__not-supported">
-				{
-					__( 'Single tickets are not yet supported on recurring events. ', 'event-tickets' )
-				}
-				<a
-					className="helper-link"
-					href="https://evnt.is/1b7a"
-					target="_blank"
-					rel="noopener noreferrer"
-				>
-					{ __( 'Read about our plans for future features', 'event-tickets' ) }
-				</a>
-			</div>
-		)
-		: null
+		<div className="tribe-editor__not-supported">
+			{
+				__( 'Single tickets are not yet supported on recurring events. ', 'event-tickets' )
+			}
+			<a
+				className="helper-link"
+				href="https://evnt.is/1b7a"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				{ __( 'Read about our plans for future features', 'event-tickets' ) }
+			</a>
+		</div>
 	);
 };
