+++ src/Tickets/Blocks/app/flexible-tickets/block-editor/components/series-pass-notice/container.js
