+++ src/Tickets/Blocks/app/flexible-tickets/block-editor/components/series-pass-notice/template.js
